+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGE LOG: Home Menu (z4m_homemenu)
 
+## Version 1.3, 2024-10-20
+- CHANGE: MOD_Z4M_HOMEMENU_COLOR_SCHEME PHP constant to customize the colors of
+  the home menu.
+- BUG FIXING: the home menu items were not focusable.
+
 ## Version 1.2, 2024-08-02
 - CHANGE: new MOD_Z4M_HOMEMENU_EXCLUDED_VIEW PHP constant to customize the name
   of the view to exclude from the home.